Implement a weekly summary view, allowing a user to easily view the average summary of macronutrient intake over the logged days over the week
Implementation did not use backend, instead it is implemented with caching for quick local update because it makes queries for each day of the week too. As a result, clicking weekly summary and then changing and choosing a selected date will be instant.

Perhaps, ideally should have been written using a backend as the current local storage implementation can cause problems if the user changes device, adds data and then the local storage pulls the cached, incorrect data. This can be changed in the future